improve comment
