+++ Sources/PoEditorParser/SwiftCodeGenerator.swift
@@ -4,49 +4,12 @@ public protocol SwiftCodeGenerator {
     func generateCode(translations: [Translation])
 }
 
-protocol CodeKeeper<Handle> {
-    associatedtype Handle
-    func append(_ content: String)
-    func build() -> Handle
-}
-
-class StringKeeper: CodeKeeper {
-    typealias Handle = String
-    private var contents: [String] = []
-    
-    func append(_ content: String) {
-        contents.append(content)
-    }
-    
-    func build() -> String {
-        return contents.joined(separator: "\n")
-    }
-}
-
-class FileHandleKeeper: CodeKeeper {
-    typealias Handle = FileHandle
-    private let handle: FileHandle
-    
-    init(handle: FileHandle) {
-        self.handle = handle
-    }
-    
-    func append(_ content: String) {
-        handle.write(content.data(using: .utf8)!)
-    }
-    
-    func build() -> FileHandle {
-        handle
-    }
-}
-
 class StructCodeGenerator: SwiftCodeGenerator {
-    
     let keeper: any CodeKeeper
     let typeName: String
     let tableName: String?
     let outputFormat: OutputFormat
-    
+
     init(
         keeper: any CodeKeeper,
         typeName: String,
@@ -58,32 +21,32 @@ class StructCodeGenerator: SwiftCodeGenerator {
         self.tableName = tableName
         self.outputFormat = outputFormat
     }
-    
+
     func generateCode(translations: [Translation]) {
         // File Header
         keeper.append(POEConstants.fileHeader)
         keeper.append(POEConstants.methodOrVariableSeparator)
-        
+
         // File contents
         keeper.append(
-        """
+            """
         public struct \(typeName): Hashable, Equatable {
             public let key: String
-            public let parameters: [String: String] 
-        
+            public let parameters: [String: String]
+
             public init(
-                key: String, 
+                key: String,
                 parameters: [String: String] = [:]
             ) {
                 self.key = key
                 self.parameters = parameters
             }
-            
+
             public func hash(into hasher: inout Hasher) {
                 hasher.combine(key)
                 hasher.combine(parameters)
             }
-        
+
             public func parsed(value: String) -> String {
                 var result = value
                 for (placeholder, replacement) in parameters {
@@ -94,34 +57,33 @@ class StructCodeGenerator: SwiftCodeGenerator {
         }
         """
         )
-        
+
         keeper.append(POEConstants.methodOrVariableSeparator)
         keeper.append(POEConstants.methodOrVariableSeparator)
         keeper.append("public extension \(typeName) {")
         keeper.append(POEConstants.methodOrVariableSeparator)
-        
+
         for (index, translation) in translations.enumerated() {
             keeper.append(translation.swiftStaticFuncCode)
             if index < translations.count - 1 {
                 keeper.append(POEConstants.methodOrVariableSeparator)
             }
         }
-        
+
         keeper.append(POEConstants.methodOrVariableSeparator)
         keeper.append("}") // Close extension
-        
+
         // File close
         keeper.append(POEConstants.fileFooter)
     }
 }
 
 class EnumCodeGenerator: SwiftCodeGenerator {
-    
     let keeper: any CodeKeeper
     let typeName: String
     let tableName: String?
     let outputFormat: OutputFormat
-    
+
     init(
         keeper: any CodeKeeper,
         typeName: String,
@@ -133,11 +95,11 @@ class EnumCodeGenerator: SwiftCodeGenerator {
         self.tableName = tableName
         self.outputFormat = outputFormat
     }
-    
+
     func generateCode(translations: [Translation]) {
         // File Header
         keeper.append(POEConstants.fileHeader)
-        
+
         // File contents
         keeper.append(POEConstants.literalsEnumHeader(keysName: typeName))
         for (index, translation) in translations.enumerated() {
@@ -148,7 +110,7 @@ class EnumCodeGenerator: SwiftCodeGenerator {
         }
         keeper.append(POEConstants.methodOrVariableSeparator)
         keeper.append(POEConstants.methodOrVariableSeparator)
-        
+
         keeper.append(POEConstants.literalsEnumValueFuncStart)
         for translation in translations {
             keeper.append(translation.swiftEnumCaseForValue)
@@ -158,9 +120,9 @@ class EnumCodeGenerator: SwiftCodeGenerator {
             keeper.append(POEConstants.literalsEnumDefaultCase)
         }
         keeper.append(POEConstants.literalsEnumValueFuncEnd)
-        
+
         keeper.append(POEConstants.methodOrVariableSeparator)
-        
+
         keeper.append(POEConstants.literalsEnumStringKeyStart)
         for translation in translations {
             keeper.append(translation.swiftEnumCaseForKey)
@@ -170,9 +132,9 @@ class EnumCodeGenerator: SwiftCodeGenerator {
             keeper.append(POEConstants.literalsEnumDefaultCase)
         }
         keeper.append(POEConstants.literalsEnumStringKeyEnd)
-        
+
         keeper.append(POEConstants.literalsEnumFooter)
-        
+
         // Filoe close
         keeper.append(POEConstants.fileFooter)
     }
@@ -183,7 +145,7 @@ public class StringCodeGenerator: SwiftCodeGenerator {
     let typeName: String
     let tableName: String?
     let outputFormat: OutputFormat
-    
+
     var codeGenerator: any SwiftCodeGenerator {
         switch outputFormat {
         case .struct:
@@ -193,6 +155,7 @@ public class StringCodeGenerator: SwiftCodeGenerator {
                 tableName: tableName,
                 outputFormat: outputFormat
             )
+
         case .enum:
             return EnumCodeGenerator(
                 keeper: keeper,
@@ -220,12 +183,11 @@ public class StringCodeGenerator: SwiftCodeGenerator {
 }
 
 public class FileCodeGenerator: SwiftCodeGenerator {
-    
     let keeper: any CodeKeeper
     let typeName: String
     let tableName: String?
     let outputFormat: OutputFormat
-    
+
     var codeGenerator: any SwiftCodeGenerator {
         switch outputFormat {
         case .struct:
@@ -235,6 +197,7 @@ public class FileCodeGenerator: SwiftCodeGenerator {
                 tableName: tableName,
                 outputFormat: outputFormat
             )
+
         case .enum:
             return EnumCodeGenerator(
                 keeper: keeper,
@@ -244,7 +207,7 @@ public class FileCodeGenerator: SwiftCodeGenerator {
             )
         }
     }
-    
+
     public init(
         fileHandle: FileHandle,
         typeName: String,
@@ -256,7 +219,7 @@ public class FileCodeGenerator: SwiftCodeGenerator {
         self.tableName = tableName
         self.outputFormat = outputFormat
     }
-    
+
     public func generateCode(translations: [Translation]) {
         codeGenerator.generateCode(translations: translations)
     }
